executionspec 383 types and a wrapper
