+++ app/ts/utils/wire-types.ts
@@ -1,6 +1,5 @@
 import * as funtypes from 'funtypes'
 import { UnionToIntersection } from './typescript.js'
-import { areEqual } from './typed-arrays.js'
 
 const BigIntParser: funtypes.ParsedValue<funtypes.String, bigint>['config'] = {
 	parse: value => {
@@ -230,13 +229,17 @@ export const EthereumUnsignedTransaction1559 = funtypes.Intersect(
 export type EthereumUnsignedTransaction = funtypes.Static<typeof EthereumUnsignedTransaction>
 export const EthereumUnsignedTransaction = funtypes.Union(EthereumUnsignedTransactionLegacy, EthereumUnsignedTransaction2930, EthereumUnsignedTransaction1559)
 
-export type EthereumTransaction2930And1559Signature = funtypes.Static<typeof EthereumTransaction2930And1559Signature>
-export const EthereumTransaction2930And1559Signature = funtypes.ReadonlyObject({
-	r: EthereumQuantity,
-	s: EthereumQuantity,
-	hash: EthereumBytes32,
-	yParity: funtypes.Union(funtypes.Literal('0x0').withParser(LiteralConverterParserFactory('0x0', 'even' as const)), funtypes.Literal('0x1').withParser(LiteralConverterParserFactory('0x1', 'odd' as const))),
-})
+export const EthereumTransaction2930And1559Signature = funtypes.Intersect(
+	funtypes.ReadonlyObject({
+		r: EthereumQuantity,
+		s: EthereumQuantity,
+		hash: EthereumBytes32,
+	}),
+	funtypes.Union(
+		funtypes.ReadonlyObject({ yParity: funtypes.Union(funtypes.Literal('0x0').withParser(LiteralConverterParserFactory('0x0', 'even' as const)), funtypes.Literal('0x1').withParser(LiteralConverterParserFactory('0x1', 'odd' as const))) }),
+		funtypes.ReadonlyObject({ v: EthereumQuantity }),
+	)
+)
 
 export type EthereumTransactionLegacySignature = funtypes.Static<typeof EthereumTransactionLegacySignature>
 export const EthereumTransactionLegacySignature = funtypes.Intersect(
@@ -294,95 +297,58 @@ export const EthereumSignedTransactionWithBlockData = funtypes.Intersect(
 	})
 )
 
-export type EthereumBlockHeader = funtypes.Static<typeof EthereumBlockHeader>
-export const EthereumBlockHeader = funtypes.ReadonlyObject({
-	author: EthereumAddress,
-	difficulty: EthereumQuantity,
-	extraData: EthereumData,
-	gasLimit: EthereumQuantity,
-	gasUsed: EthereumQuantity,
-	hash: EthereumBytes32,
-	logsBloom: EthereumBytes256,
-	miner: EthereumAddress,
-	mixHash: EthereumBytes32,
-	nonce: EthereumBytes16,
-	number: EthereumQuantity,
-	parentHash: EthereumBytes32,
-	receiptsRoot: EthereumBytes32,
-	sha3Uncles: EthereumBytes32,
-	stateRoot: EthereumBytes32,
-	timestamp: EthereumTimestamp,
-	size: EthereumQuantity,
-	totalDifficulty: EthereumQuantity,
-	transactions: funtypes.ReadonlyArray(EthereumSignedTransaction),
-	uncles: funtypes.ReadonlyArray(EthereumBytes32),
-	baseFeePerGas: funtypes.Union(EthereumQuantity, funtypes.Undefined),
-	transactionsRoot: EthereumBytes32
-}).asReadonly()
+export type EthereumWithdrawal = funtypes.Static<typeof EthereumWithdrawal>
+export const EthereumWithdrawal = funtypes.ReadonlyObject({
+	index: EthereumQuantity,
+	validatorIndex: EthereumQuantity,
+	address: EthereumAddress,
+	amount: EthereumQuantity,
+})
 
-export type EthGetStorageAtResponse = funtypes.Static<typeof EthGetStorageAtResponse>
-export const EthGetStorageAtResponse = funtypes.Union(
-	EthereumBytes32,
-	funtypes.String.withParser({ parse: x => x === '0x' ? { success: true, value: null } : { success: false, message: `eth_getStorageAt didn't return 32 bytes of data nor 0x.` } }),
-)
-
-export type EthGetLogsRequest = funtypes.Static<typeof EthGetLogsRequest>
-export const EthGetLogsRequest = funtypes.Intersect(
-	funtypes.Union(
-		funtypes.ReadonlyObject({ blockHash: EthereumBytes32 }).asReadonly(),
-		funtypes.Partial({ fromBlock: EthereumBlockTag, toBlock: EthereumBlockTag }).asReadonly(),
-	),
-	funtypes.Partial({
-		address: funtypes.Union(EthereumAddress, funtypes.ReadonlyArray(EthereumAddress)),
-		topics: funtypes.ReadonlyArray(funtypes.Union(EthereumBytes32, funtypes.ReadonlyArray(EthereumBytes32), funtypes.Null)),
-	}).asReadonly()
-)
-
-export type EthGetLogsResponse = funtypes.Static<typeof EthGetLogsResponse>
-export const EthGetLogsResponse = funtypes.ReadonlyArray(
+type EthereumBlockHeaderWithoutTransactions = funtypes.Static<typeof EthereumBlockHeaderWithoutTransactions>
+const EthereumBlockHeaderWithoutTransactions = funtypes.Intersect(
+	funtypes.MutablePartial({
+		author: EthereumAddress,
+	}),
 	funtypes.ReadonlyObject({
-		removed: funtypes.Boolean,
-		logIndex: funtypes.Union(EthereumQuantity, funtypes.Null),
-		transactionIndex: funtypes.Union(EthereumQuantity, funtypes.Null),
-		transactionHash: funtypes.Union(EthereumBytes32, funtypes.Null),
-		blockHash: funtypes.Union(EthereumBytes32, funtypes.Null),
-		blockNumber: funtypes.Union(EthereumQuantity, funtypes.Null),
-		address: EthereumAddress,
-		data: EthereumInput,
-		topics: funtypes.ReadonlyArray(EthereumBytes32),
-	}).asReadonly()
-)
-
-export type EthTransactionReceiptResponse = funtypes.Static<typeof EthTransactionReceiptResponse>
-export const EthTransactionReceiptResponse = funtypes.Union(
-	funtypes.Null,
-	funtypes.ReadonlyObject({
-		type: funtypes.Union(
-			funtypes.Union(funtypes.Literal('0x0').withParser(LiteralConverterParserFactory('0x0', 'legacy' as const)), funtypes.Literal(undefined).withParser(LiteralConverterParserFactory(undefined, 'legacy' as const))),
-			funtypes.Literal('0x0').withParser(LiteralConverterParserFactory('0x0', 'legacy' as const)),
-			funtypes.Literal('0x1').withParser(LiteralConverterParserFactory('0x1', '2930' as const)),
-			funtypes.Literal('0x2').withParser(LiteralConverterParserFactory('0x2', '1559' as const)),
-		),
-		blockHash: EthereumBytes32,
-		blockNumber: EthereumQuantity,
-		transactionHash: EthereumBytes32,
-		transactionIndex: EthereumQuantity,
-		contractAddress: funtypes.Union(funtypes.Null, EthereumAddress),
-		cumulativeGasUsed: EthereumQuantity,
+		difficulty: EthereumQuantity,
+		extraData: EthereumData,
+		gasLimit: EthereumQuantity,
 		gasUsed: EthereumQuantity,
-		effectiveGasPrice: EthereumQuantity,
-		from: EthereumAddress,
-		to: funtypes.Union(funtypes.Null, EthereumAddress),
-		logs: EthGetLogsResponse,
+		hash: EthereumBytes32,
 		logsBloom: EthereumBytes256,
-		status: funtypes.Union(
-			funtypes.Literal('0x0').withParser(LiteralConverterParserFactory('0x0', 'failure' as const)),
-			funtypes.Literal('0x1').withParser(LiteralConverterParserFactory('0x1', 'success' as const)),
-		),
-	}).asReadonly()
+		miner: EthereumAddress,
+		mixHash: EthereumBytes32,
+		nonce: EthereumBytes16,
+		number: EthereumQuantity,
+		parentHash: EthereumBytes32,
+		receiptsRoot: EthereumBytes32,
+		sha3Uncles: EthereumBytes32,
+		stateRoot: EthereumBytes32,
+		timestamp: EthereumTimestamp,
+		size: EthereumQuantity,
+		totalDifficulty: EthereumQuantity,
+		uncles: funtypes.ReadonlyArray(EthereumBytes32),
+		baseFeePerGas: funtypes.Union(EthereumQuantity, funtypes.Undefined),
+		transactionsRoot: EthereumBytes32,
+		withdrawalsRoot: EthereumBytes32,
+		withdrawals: funtypes.ReadonlyArray(EthereumWithdrawal),
+	})
 )
 
-const RevertErrorParser: funtypes.ParsedValue<funtypes.String, string>['config'] = {
+export type EthereumBlockHeaderWithTransactionHashes = funtypes.Static<typeof EthereumBlockHeaderWithTransactionHashes>
+export const EthereumBlockHeaderWithTransactionHashes = funtypes.Intersect(
+	EthereumBlockHeaderWithoutTransactions,
+	funtypes.ReadonlyObject({ transactions: funtypes.ReadonlyArray(EthereumBytes32) })
+)
+
+export type EthereumBlockHeader = funtypes.Static<typeof EthereumBlockHeader>
+export const EthereumBlockHeader = funtypes.Intersect(
+	EthereumBlockHeaderWithoutTransactions,
+	funtypes.ReadonlyObject({ transactions: funtypes.ReadonlyArray(EthereumSignedTransaction) })
+)
+
+export const RevertErrorParser: funtypes.ParsedValue<funtypes.String, string>['config'] = {
 	parse: value => {
 		if (!value.startsWith('Reverted ')) return { success: true, value }
 		const parseResult = BytesParser.parse(value.slice('Reverted '.length))
@@ -397,52 +363,6 @@ const RevertErrorParser: funtypes.ParsedValue<funtypes.String, string>['config']
 		return { success: true, value: `Reverted ${serializationResult.value}` }
 	},
 }
-
-export type MulticallRequestParameters = funtypes.Static<typeof MulticallRequestParameters>
-export const MulticallRequestParameters = funtypes.ReadonlyTuple(
-	EthereumQuantity, // block number
-	EthereumAddress, // miner
-	funtypes.ReadonlyArray(EthereumUnsignedTransaction),
-)
-
-export type MulticallResponseEventLog = funtypes.Static<typeof MulticallResponseEventLog>
-export const MulticallResponseEventLog =  funtypes.ReadonlyObject({
-	loggersAddress: EthereumAddress,
-	data: EthereumInput,
-	topics: funtypes.ReadonlyArray(EthereumBytes32),
-}).asReadonly()
-
-export type MulticallResponseEventLogs = funtypes.Static<typeof MulticallResponseEventLogs>
-export const MulticallResponseEventLogs = funtypes.ReadonlyArray(MulticallResponseEventLog)
-
-export type EthBalanceChanges = funtypes.Static<typeof EthBalanceChanges>
-export const EthBalanceChanges = funtypes.ReadonlyArray(
-	funtypes.ReadonlyObject({
-		address: EthereumAddress,
-		before: EthereumQuantity,
-		after: EthereumQuantity,
-	}).asReadonly()
-)
-
-export type SingleMulticallResponse = funtypes.Static<typeof SingleMulticallResponse>
-export const SingleMulticallResponse = funtypes.Union(
-	funtypes.ReadonlyObject({
-		statusCode: funtypes.Literal(1).withParser(LiteralConverterParserFactory(1, 'success' as const)),
-		gasSpent: EthereumQuantity,
-		returnValue: EthereumData,
-		events: MulticallResponseEventLogs,
-		balanceChanges: EthBalanceChanges,
-	}).asReadonly(),
-	funtypes.ReadonlyObject({
-		statusCode: funtypes.Literal(0).withParser(LiteralConverterParserFactory(0, 'failure' as const)),
-		gasSpent: EthereumQuantity,
-		error: funtypes.String.withParser(RevertErrorParser),
-		returnValue: EthereumData,
-	}).asReadonly(),
-)
-
-export type MulticallResponse = funtypes.Static<typeof MulticallResponse>
-export const MulticallResponse = funtypes.ReadonlyArray(SingleMulticallResponse)
 
 //
 // Token Lists
@@ -490,371 +410,3 @@ export type ToWireType<T> =
 	: T extends funtypes.ParsedValue<infer U, infer _> ? ToWireType<U>
 	: T extends funtypes.Codec<infer U> ? U
 	: never
-
-
-export type DappRequestTransaction = funtypes.Static<typeof DappRequestTransaction>
-export const DappRequestTransaction = funtypes.ReadonlyPartial({
-	from: EthereumAddress,
-	gas: EthereumQuantity,
-	value: EthereumQuantity,
-	to: funtypes.Union(EthereumAddress, funtypes.Null),
-	gasPrice: EthereumQuantity,
-	maxPriorityFeePerGas: funtypes.Union(EthereumQuantity, funtypes.Null), // etherscan sets this field to null, remove this if etherscan fixes this
-	maxFeePerGas: funtypes.Union(EthereumQuantity, funtypes.Null), // etherscan sets this field to null, remove this if etherscan fixes this
-	data: EthereumData,
-	input: EthereumData,
-}).withConstraint((dappRequestTransaction) => {
-	if (dappRequestTransaction.input !== undefined && dappRequestTransaction.data !== undefined) {
-		if (typeof dappRequestTransaction.input === 'string') {
-			// workaround for https://github.com/ForbesLindesay/funtypes/issues/62
-			return dappRequestTransaction.input === dappRequestTransaction.data
-		}
-		return areEqual(dappRequestTransaction.input, dappRequestTransaction.data)
-	}
-	return true
-})
-.withConstraint((x) => {
-	if (x.gasPrice !== undefined) {
-		return x.maxPriorityFeePerGas === undefined && x.maxFeePerGas === undefined
-	} else if (x.maxPriorityFeePerGas !== undefined) {
-		return x.maxFeePerGas !== undefined && x.gasPrice === undefined
-	} else if (x.maxFeePerGas !== undefined) {
-		return x.maxPriorityFeePerGas !== undefined && x.gasPrice === undefined
-	} else {
-		return true
-	}
-})
-
-export type EthereumBlockHeaderWithTransactionHashes = funtypes.Static<typeof EthereumBlockHeaderWithTransactionHashes>
-export const EthereumBlockHeaderWithTransactionHashes = funtypes.ReadonlyObject({
-	author: EthereumAddress,
-	difficulty: EthereumQuantity,
-	extraData: EthereumData,
-	gasLimit: EthereumQuantity,
-	gasUsed: EthereumQuantity,
-	hash: EthereumBytes32,
-	logsBloom: EthereumBytes256,
-	miner: EthereumAddress,
-	mixHash: EthereumBytes32,
-	nonce: EthereumBytes16,
-	number: EthereumQuantity,
-	parentHash: EthereumBytes32,
-	receiptsRoot: EthereumBytes32,
-	sha3Uncles: EthereumBytes32,
-	stateRoot: EthereumBytes32,
-	timestamp: EthereumTimestamp,
-	size: EthereumQuantity,
-	totalDifficulty: EthereumQuantity,
-	transactions: funtypes.ReadonlyArray(EthereumBytes32),
-	uncles: funtypes.ReadonlyArray(EthereumBytes32),
-	baseFeePerGas: funtypes.Union(EthereumQuantity, funtypes.Undefined),
-	transactionsRoot: EthereumBytes32
-}).asReadonly()
-
-export type GetBlockReturn = funtypes.Static<typeof GetBlockReturn>
-export const GetBlockReturn = funtypes.Union(EthereumBlockHeader, EthereumBlockHeaderWithTransactionHashes)
-
-export const NewHeadsSubscriptionData = funtypes.ReadonlyObject({
-	subscription: funtypes.String,
-	result: EthereumBlockHeaderWithTransactionHashes
-}).asReadonly()
-
-export type JsonRpcNewHeadsNotification = funtypes.Static<typeof JsonRpcNewHeadsNotification>
-export const JsonRpcNewHeadsNotification = funtypes.ReadonlyObject({
-	jsonrpc: funtypes.Literal('2.0'),
-	method: funtypes.String,
-	params: NewHeadsSubscriptionData
-}).asReadonly()
-
-export type JsonSubscriptionNotification = funtypes.Static<typeof JsonSubscriptionNotification>
-export const JsonSubscriptionNotification = funtypes.ReadonlyObject({
-	jsonrpc: funtypes.Literal('2.0'),
-	method: funtypes.Literal('eth_subscription'),
-	params: funtypes.ReadonlyObject({
-		result: funtypes.Union(EthereumBlockHeader, EthereumBytes32),
-		subscription: funtypes.String
-	}).asReadonly()
-}).asReadonly()
-
-export type JsonRpcSuccessResponse = funtypes.Static<typeof JsonRpcSuccessResponse>
-export const JsonRpcSuccessResponse = funtypes.ReadonlyObject({
-	jsonrpc: funtypes.Literal('2.0'),
-	id: funtypes.Union(funtypes.String, funtypes.Number),
-	result: funtypes.Unknown,
-}).asReadonly()
-
-export type JsonRpcErrorResponse = funtypes.Static<typeof JsonRpcErrorResponse>
-export const JsonRpcErrorResponse = funtypes.ReadonlyObject({
-	jsonrpc: funtypes.Literal('2.0'),
-	id: funtypes.Union(funtypes.String, funtypes.Number),
-	error: funtypes.ReadonlyObject({
-		code: funtypes.Number,
-		message: funtypes.String,
-		data: funtypes.Unknown,
-	}).asReadonly(),
-}).asReadonly()
-
-
-export type JsonRpcNotification = funtypes.Static<typeof JsonRpcNotification>
-export const JsonRpcNotification = funtypes.Union(JsonRpcNewHeadsNotification, JsonSubscriptionNotification)
-
-export type JsonRpcRequest = funtypes.Static<typeof JsonRpcRequest>
-export const JsonRpcRequest = funtypes.ReadonlyObject({
-	jsonrpc: funtypes.Literal('2.0'),
-	id: funtypes.Union(funtypes.String, funtypes.Number),
-	method: funtypes.String,
-	params: funtypes.Union(funtypes.ReadonlyArray(funtypes.Unknown), funtypes.Literal(undefined).withParser(LiteralConverterParserFactory(undefined, [])))
-})
-
-export type JsonRpcResponse = funtypes.Static<typeof JsonRpcResponse>
-export const JsonRpcResponse = funtypes.Union(JsonRpcErrorResponse, JsonRpcSuccessResponse)
-
-export type JsonRpcMessage = funtypes.Static<typeof JsonRpcMessage>
-export const JsonRpcMessage = funtypes.Union(JsonRpcResponse, JsonRpcNotification, JsonRpcRequest)
-
-export type TransactionByHashParams = funtypes.Static<typeof TransactionByHashParams>
-export const TransactionByHashParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_getTransactionByHash'),
-	params: funtypes.ReadonlyTuple(EthereumBytes32)
-})
-
-export type SendTransactionParams = funtypes.Static<typeof SendTransactionParams>
-export const SendTransactionParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_sendTransaction'),
-	params: funtypes.ReadonlyTuple(DappRequestTransaction)
-})
-
-export type SendRawTransaction = funtypes.Static<typeof SendRawTransaction>
-export const SendRawTransaction = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_sendRawTransaction'),
-	params: funtypes.ReadonlyTuple(EthereumData),
-})
-
-export type EthereumAccountsReply = funtypes.Static<typeof EthereumAccountsReply>
-export const EthereumAccountsReply = funtypes.ReadonlyTuple(funtypes.ReadonlyArray(EthereumAddress), funtypes.Boolean)
-
-export type EthereumChainReply = funtypes.Static<typeof EthereumChainReply>
-export const EthereumChainReply = funtypes.ReadonlyArray(EthereumQuantity)
-
-export type TransactionReceiptParams = funtypes.Static<typeof TransactionReceiptParams>
-export const TransactionReceiptParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_getTransactionReceipt'),
-	params: funtypes.ReadonlyTuple(EthereumBytes32)
-})
-
-export type EstimateGasParams = funtypes.Static<typeof EstimateGasParams>
-export const EstimateGasParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_estimateGas'),
-	params: funtypes.Union(funtypes.ReadonlyTuple(DappRequestTransaction), funtypes.ReadonlyTuple(DappRequestTransaction, EthereumBlockTag))
-})
-
-export type EthCallParams = funtypes.Static<typeof EthCallParams>
-export const EthCallParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_call'),
-	params: funtypes.ReadonlyTuple(
-		DappRequestTransaction,
-		EthereumBlockTag
-	)
-}).asReadonly()
-
-export type EthGetLogsParams = funtypes.Static<typeof EthGetLogsParams>
-export const EthGetLogsParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_getLogs'),
-	params: funtypes.ReadonlyTuple(EthGetLogsRequest)
-}).asReadonly()
-
-export type EthBalanceParams = funtypes.Static<typeof EthBalanceParams>
-export const EthBalanceParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_getBalance'),
-	params: funtypes.ReadonlyTuple(EthereumAddress, EthereumBlockTag)
-})
-
-export type EthBlockByNumberParams = funtypes.Static<typeof EthBlockByNumberParams>
-export const EthBlockByNumberParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_getBlockByNumber'),
-	params: funtypes.ReadonlyTuple(EthereumBlockTag, funtypes.Boolean)
-})
-
-export type EthSubscribeParams = funtypes.Static<typeof EthSubscribeParams>
-export const EthSubscribeParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_subscribe'),
-	params: funtypes.ReadonlyTuple(funtypes.Union(funtypes.Literal('newHeads'), funtypes.Literal('logs'), funtypes.Literal('newPendingTransactions'), funtypes.Literal('syncing')))
-})
-
-export type EthUnSubscribeParams = funtypes.Static<typeof EthUnSubscribeParams>
-export const EthUnSubscribeParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_unsubscribe'),
-	params: funtypes.ReadonlyTuple(funtypes.String)
-})
-
-export type EthGetStorageAtParams = funtypes.Static<typeof EthGetStorageAtParams>
-export const EthGetStorageAtParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_getStorageAt'),
-	params: funtypes.ReadonlyTuple(EthereumAddress, EthereumQuantity, EthereumBlockTag)
-})
-
-export const EthSubscriptionResponse = funtypes.String
-export type EthSubscriptionResponse = funtypes.Static<typeof EthSubscriptionResponse>
-
-export type PersonalSignParams = funtypes.Static<typeof PersonalSignParams>
-export const PersonalSignParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('personal_sign'),
-	params: funtypes.Union(
-		funtypes.ReadonlyTuple(funtypes.String, EthereumAddress, funtypes.Union(funtypes.String, funtypes.Undefined, funtypes.Null)), // message, account, password
-		funtypes.ReadonlyTuple(funtypes.String, EthereumAddress) // message, account
-	)
-})
-
-type typeJSONEncodeable = string | number | boolean | { [x: string]: typeJSONEncodeable | undefined } | ReadonlyArray<typeJSONEncodeable>
-export type JSONEncodeable = funtypes.Static<typeof JSONEncodeable>
-export const JSONEncodeable: funtypes.Runtype<typeJSONEncodeable> = funtypes.Lazy(() => funtypes.Union(
-	funtypes.String,
-	funtypes.Boolean,
-	funtypes.Number,
-	funtypes.ReadonlyArray(JSONEncodeable),
-	funtypes.ReadonlyRecord(funtypes.String, JSONEncodeable),
-))
-
-export type JSONEncodeableObject = funtypes.Static<typeof JSONEncodeableObject>
-export const JSONEncodeableObject = funtypes.ReadonlyRecord(funtypes.String, JSONEncodeable)
-
-export type JSONEncodeableObjectOrArray = funtypes.Static<typeof JSONEncodeableObjectOrArray>
-export const JSONEncodeableObjectOrArray = funtypes.Union(funtypes.ReadonlyArray(JSONEncodeable), funtypes.ReadonlyRecord(funtypes.String, JSONEncodeable))
-
-export type EIP712MessageUnderlying = funtypes.Static<typeof EIP712MessageUnderlying>
-export const EIP712MessageUnderlying = funtypes.ReadonlyObject({
-	types: funtypes.Record(funtypes.String, funtypes.ReadonlyArray(
-		funtypes.ReadonlyObject({
-			name: funtypes.String,
-			type: funtypes.String,
-		})
-	)),
-	primaryType: funtypes.String,
-	domain: JSONEncodeableObject,
-	message: JSONEncodeableObject,
-})
-
-export function isJSON(text: string){
-	if (typeof text !== 'string') return false
-	try {
-		JSON.parse(text)
-		return true
-	}
-	catch (error) {
-		return false
-	}
-}
-
-const EIP712MessageParser: funtypes.ParsedValue<funtypes.String, EIP712MessageUnderlying>['config'] = {
-	parse: value => {
-		if (!isJSON(value) || !EIP712MessageUnderlying.test(JSON.parse(value))) return { success: false, message: `${ value } is not EIP712 message` }
-		else return { success: true, value: EIP712MessageUnderlying.parse(JSON.parse(value)) }
-	},
-	serialize: value => {
-		if (!EIP712MessageUnderlying.test(value)) return { success: false, message: `${ value } is not a EIP712 message.`}
-		return { success: true, value: JSON.stringify(EIP712MessageUnderlying.serialize(value)) }
-	},
-}
-
-export type EIP712Message = funtypes.Static<typeof EIP712Message>
-export const EIP712Message = funtypes.String.withParser(EIP712MessageParser)
-
-export type OldSignTypedDataParams = funtypes.Static<typeof OldSignTypedDataParams>
-export const OldSignTypedDataParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_signTypedData'),
-	params: funtypes.ReadonlyTuple(funtypes.ReadonlyArray(
-		funtypes.ReadonlyObject({
-			name: funtypes.String,
-			type: funtypes.String,
-		})
-	), EthereumAddress),
-})
-
-export type SignTypedDataParams = funtypes.Static<typeof SignTypedDataParams>
-export const SignTypedDataParams = funtypes.ReadonlyObject({
-	method: funtypes.Union(
-		funtypes.Literal('eth_signTypedData_v1'),
-		funtypes.Literal('eth_signTypedData_v2'),
-		funtypes.Literal('eth_signTypedData_v3'),
-		funtypes.Literal('eth_signTypedData_v4'),
-	),
-	params: funtypes.ReadonlyTuple(EthereumAddress, EIP712Message), // address that will sign the message, typed data
-})
-export type SwitchEthereumChainParams = funtypes.Static<typeof SwitchEthereumChainParams>
-export const SwitchEthereumChainParams = funtypes.ReadonlyObject({
-	method: funtypes.Literal('wallet_switchEthereumChain'),
-	params: funtypes.Tuple(funtypes.ReadonlyObject({ chainId: EthereumQuantity }).asReadonly()),
-}).asReadonly()
-
-export type GetCode = funtypes.Static<typeof GetCode>
-export const GetCode = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_getCode'),
-	params: funtypes.ReadonlyTuple(EthereumAddress, EthereumBlockTag)
-}).asReadonly()
-
-export type RequestPermissions = funtypes.Static<typeof RequestPermissions>
-export const RequestPermissions = funtypes.ReadonlyObject({
-	method: funtypes.Literal('wallet_requestPermissions'),
-	params: funtypes.ReadonlyTuple( funtypes.ReadonlyObject({ eth_accounts: funtypes.ReadonlyObject({ }) }) )
-}).asReadonly()
-
-export type GetTransactionCount = funtypes.Static<typeof GetTransactionCount>
-export const GetTransactionCount = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_getTransactionCount'),
-	params: funtypes.ReadonlyTuple(EthereumAddress, EthereumBlockTag)
-}).asReadonly()
-
-export type EthSign = funtypes.Static<typeof EthSign>
-export const EthSign = funtypes.ReadonlyObject({
-	method: funtypes.Literal('eth_sign'),
-	params: funtypes.ReadonlyTuple(EthereumAddress, funtypes.String),
-}).asReadonly()
-
-export type GetSimulationStackReply = funtypes.Static<typeof GetSimulationStackReply>
-export const GetSimulationStackReply = funtypes.ReadonlyArray(funtypes.Intersect(
-	EthereumUnsignedTransaction,
-	SingleMulticallResponse,
-	funtypes.ReadonlyObject({
-		realizedGasPrice: EthereumQuantity,
-		gasLimit: EthereumQuantity,
-	}).asReadonly(),
-))
-
-export type GetSimulationStack = funtypes.Static<typeof GetSimulationStack>
-export const GetSimulationStack = funtypes.ReadonlyObject({
-	method: funtypes.Literal('interceptor_getSimulationStack'),
-	params: funtypes.ReadonlyTuple(funtypes.Literal('1.0.0')),
-}).asReadonly()
-
-export type EthereumJsonRpcRequest = funtypes.Static<typeof EthereumJsonRpcRequest>
-export const EthereumJsonRpcRequest = funtypes.Union(
-	EthBlockByNumberParams,
-	EthBalanceParams,
-	EstimateGasParams,
-	TransactionByHashParams,
-	TransactionReceiptParams,
-	SendTransactionParams,
-	SendRawTransaction,
-	EthCallParams,
-	EthSubscribeParams,
-	EthUnSubscribeParams,
-	funtypes.ReadonlyObject({ method: funtypes.Literal('eth_blockNumber') }),
-	funtypes.ReadonlyObject({ method: funtypes.Literal('eth_chainId') }),
-	funtypes.ReadonlyObject({ method: funtypes.Literal('net_version') }),
-	GetCode,
-	PersonalSignParams,
-	OldSignTypedDataParams,
-	SignTypedDataParams,
-	SwitchEthereumChainParams,
-	RequestPermissions,
-	funtypes.ReadonlyObject({ method: funtypes.Literal('wallet_getPermissions') }),
-	funtypes.ReadonlyObject({ method: funtypes.Literal('eth_accounts') }),
-	funtypes.ReadonlyObject({ method: funtypes.Literal('eth_requestAccounts') }),
-	funtypes.ReadonlyObject({ method: funtypes.Literal('eth_gasPrice') }),
-	GetTransactionCount,
-	GetSimulationStack,
-	funtypes.ReadonlyObject({ method: funtypes.Literal('eth_multicall'), params: MulticallRequestParameters }),
-	EthGetStorageAtParams,
-	EthGetLogsParams,
-	EthSign,
-)
