+++ test/nethermindRPCResponses.ts
@@ -28,7 +28,9 @@ export const eth_getBlockByNumber_goerli_8443561_false = `
             "0xb47867ca1e305b8bdcab14ebb8dffaf77bd4e52016de1d47a41a0619018d8b74"
         ],
         "transactionsRoot": "0x326f75273fb880b00edec6d45dcfa0f9f75b86600ddfefba16998fce49dcad14",
-        "uncles": []
+        "uncles": [],
+        "withdrawalsRoot": "0x326f75273fb880b00edec6d45dcfa0f9f75b86600ddfefba16998fce49dcad14",
+        "withdrawals": []
     },
     "id": 2
 }`
